Change maximum size of application messages
- len of application messages was an unsigned short. It is now an int to fit the calculations which were done in Simatic's PhD thesis.

+++ include/trains.h
@@ -52,7 +52,7 @@ extern address my_address;
  * The @a payload contains the information carried by the protocol for the application.
  */
 typedef struct{
-  unsigned short len;                /**< Length of whole message */
+  int            len;                /**< Length of whole message */
   char           payload[];          /**< Payload (i.e. contents) of the message */
 } message;
